Add recovery mode
The database is now able to recover its state from existing files!
Also added some advanced logging capabilites; the DB server now
writes to a file rather than the console.

+++ db/log_reader.cc
@@ -7,7 +7,7 @@
 
 namespace mdb {
 
-LogReader::LogReader(int log_number, const Options& options)
+LogReader::LogReader(size_t log_number, const Options& options)
     : LogReader(
           options.env->MakeReadOnlyIO(util::LogFileName(options, log_number))) {
 }
