+++ db/log_reader.h
@@ -12,7 +12,7 @@ namespace mdb {
 
 class LogReader {
  public:
-  LogReader(int log_number, const Options& options);
+  LogReader(size_t log_number, const Options& options);
   explicit LogReader(std::unique_ptr<ReadOnlyIO>&& file);
 
   MemTableT ReadMemTable();
